feat: separate fetch brand data into custom hook
